backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
python-dotenv
requests
cryptography
pytest
httpx
